migrated info from memory to world model where appropriate
