+++ scripts/scene_elements/commander.py
@@ -13,6 +13,7 @@ __all__ = ["Commander"]
 
 class Commander:
     def __init__(self, game: Game, type_: str):
+        # TODO - is this needed? What is it doing for us?
         self._game: Game = game
 
         self.type = type_
@@ -20,10 +21,3 @@ class Commander:
 
         data = self._game.data.commanders[type_]
         self.name = data["name"]
-
-    @property
-    def charisma_remaining(self) -> int:
-        num_units = len(self._game.memory.player_troupe.units)
-        remaining = self._game.memory.charisma - num_units
-
-        return remaining
