+++ scripts/scene_elements/projectile.py
@@ -30,7 +30,7 @@ class Projectile:
             if not self._game.world.ui.terrain.check_tile_hoverable(self.pos):
                 return False
 
-            for entity in self._game.memory.get_all_entities():
+            for entity in self._game.world.model.get_all_entities():
                 if entity.team != self.owner.team:
                     if r.collidepoint(entity.pos):
                         entity.deal_damage(self.damage, self.owner)
